feat: moved propose button to proper location
